+++ src/components/settings/settings-root-proxy/_fields/root-proxy-select/root-proxy-select-field.js
@@ -4,11 +4,13 @@ import {
   useIntl,
 } from 'react-intl';
 
+import { useStripes } from '@folio/stripes/core';
 import { Select } from '@folio/stripes/components';
 import styles from './root-proxy-select-field.css';
 
 const RootProxySelectField = ({ proxyTypes }) => {
   const intl = useIntl();
+  const stripes = useStripes();
   const label = intl.formatMessage({ id: 'ui-eholdings.settings.rootProxy.server' });
 
   if (!proxyTypes?.items?.length) {
@@ -37,6 +39,7 @@ const RootProxySelectField = ({ proxyTypes }) => {
         component={Select}
         label={label}
         aria-label={label}
+        disabled={!stripes.hasPerm('ui-eholdings.settings.root-proxy.edit')}
         data-testid="root-proxy-select-field"
       >
         {options}