+++ src/routes/settings-root-proxy-route/settings-root-proxy-route.js
@@ -1,7 +1,10 @@
 import { useEffect } from 'react';
 import PropTypes from 'prop-types';
 import ReactRouterPropTypes from 'react-router-prop-types';
-import { TitleManager } from '@folio/stripes/core';
+import {
+  TitleManager,
+  useStripes,
+} from '@folio/stripes/core';
 import { useIntl } from 'react-intl';
 
 import View from '../../components/settings/settings-root-proxy';
@@ -30,11 +33,21 @@ const SettingsRootProxyRoute = ({
   updateRootProxy,
 }) => {
   const intl = useIntl();
+  const stripes = useStripes();
+  const hasPermToView = stripes.hasPerm('ui-eholdings.settings.root-proxy.view');
+
+  if (!hasPermToView) {
+    history.push('/settings/eholdings');
+  }
 
   useEffect(() => {
+    if (!hasPermToView) {
+      return;
+    }
+
     getProxyTypes(match.params.kbId);
     getRootProxy(match.params.kbId);
-  }, [getProxyTypes, getRootProxy, match.params.kbId]);
+  }, [getProxyTypes, getRootProxy, match.params.kbId, hasPermToView]);
 
   useEffect(() => {
     if (rootProxy.isUpdated) {